--- server/package.json
@@ -1,20 +1,0 @@
-{
-  "name": "server",
-  "version": "1.0.0",
-  "description": "",
-  "main": "src/index.js",
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
-  },
-  "keywords": [],
-  "author": "",
-  "license": "ISC",
-  "dependencies": {
-    "bcryptjs": "^2.4.3",
-    "graphql-yoga": "^1.17.4",
-    "jsonwebtoken": "^8.5.0",
-    "nodemon": "^1.18.11",
-    "prisma-client-lib": "^1.28.0",
-    "request": "^2.88.0"
-  }
-}
